:high_brightness: Add helper method for MDC
